// src/types/api/service/models/BaseResponseCommonObject.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MessageDto } from './MessageDto';
import {
    MessageDtoFromJSON,
    MessageDtoFromJSONTyped,
    MessageDtoToJSON,
} from './MessageDto';

/**
 * 
 * @export
 * @interface BaseResponseCommonObject
 */
export interface BaseResponseCommonObject {
    /**
     * 
     * @type {string}
     * @memberof BaseResponseCommonObject
     */
    status?: string;
    /**
     * 
     * @type {object}
     * @memberof BaseResponseCommonObject
     */
    data?: object;
    /**
     * 
     * @type {string}
     * @memberof BaseResponseCommonObject
     */
    code?: string;
    /**
     * 
     * @type {MessageDto}
     * @memberof BaseResponseCommonObject
     */
    message?: MessageDto;
    /**
     * 
     * @type {string}
     * @memberof BaseResponseCommonObject
     */
    timestamp?: string;
}

/**
 * Check if a given object implements the BaseResponseCommonObject interface.
 */
export function instanceOfBaseResponseCommonObject(value: object): value is BaseResponseCommonObject {
    return true;
}

export function BaseResponseCommonObjectFromJSON(json: any): BaseResponseCommonObject {
    return BaseResponseCommonObjectFromJSONTyped(json, false);
}

export function BaseResponseCommonObjectFromJSONTyped(json: any, ignoreDiscriminator: boolean): BaseResponseCommonObject {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'] == null ? undefined : json['status'],
        'data': json['data'] == null ? undefined : json['data'],
        'code': json['code'] == null ? undefined : json['code'],
        'message': json['message'] == null ? undefined : MessageDtoFromJSON(json['message']),
        'timestamp': json['timestamp'] == null ? undefined : json['timestamp'],
    };
}

export function BaseResponseCommonObjectToJSON(value?: BaseResponseCommonObject | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'status': value['status'],
        'data': value['data'],
        'code': value['code'],
        'message': MessageDtoToJSON(value['message']),
        'timestamp': value['timestamp'],
    };
}