// src/types/api/service/models/GroupDtoForUser.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GroupDtoForUser
 */
export interface GroupDtoForUser {
    /**
     * 
     * @type {number}
     * @memberof GroupDtoForUser
     */
    groupId?: number;
    /**
     * 
     * @type {string}
     * @memberof GroupDtoForUser
     */
    name?: string;
    /**
     * 
     * @type {boolean}
     * @memberof GroupDtoForUser
     */
    isDefault?: boolean;
}

/**
 * Check if a given object implements the GroupDtoForUser interface.
 */
export function instanceOfGroupDtoForUser(value: object): value is GroupDtoForUser {
    return true;
}

export function GroupDtoForUserFromJSON(json: any): GroupDtoForUser {
    return GroupDtoForUserFromJSONTyped(json, false);
}

export function GroupDtoForUserFromJSONTyped(json: any, ignoreDiscriminator: boolean): GroupDtoForUser {
    if (json == null) {
        return json;
    }
    return {
        
        'groupId': json['groupId'] == null ? undefined : json['groupId'],
        'name': json['name'] == null ? undefined : json['name'],
        'isDefault': json['isDefault'] == null ? undefined : json['isDefault'],
    };
}

export function GroupDtoForUserToJSON(value?: GroupDtoForUser | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'groupId': value['groupId'],
        'name': value['name'],
        'isDefault': value['isDefault'],
    };
}